#193: move picker OS-wiring into container, extract WindowDropHandler, private MacFilePicker method, abstract ADR example
