+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/transfer/MacFilePicker.kt
@@ -21,33 +21,33 @@ internal class MacFilePicker(
         }
     }
 
-    private companion object {
-        private const val NS_MODAL_RESPONSE_OK = 1
-
-        suspend fun pickFilesAndFolders(): List<File> = withContext(Dispatchers.IO) {
-            val pool = Foundation.NSAutoreleasePool()
-            try {
-                var result: List<File> = emptyList()
-                Foundation.executeOnMainThread(withAutoreleasePool = true, waitUntilDone = true) {
-                    val panel = Foundation.invoke("NSOpenPanel", "openPanel")
-                    Foundation.invoke(panel, "setCanChooseFiles:", true)
-                    Foundation.invoke(panel, "setCanChooseDirectories:", true)
-                    Foundation.invoke(panel, "setAllowsMultipleSelection:", true)
-                    val response = Foundation.invoke(panel, "runModal")
-                    if (response.toInt() == NS_MODAL_RESPONSE_OK) {
-                        val urls = Foundation.invoke(panel, "URLs")
-                        val count = Foundation.invoke(urls, "count").toInt()
-                        result = (0 until count).mapNotNull { i ->
-                            val url = Foundation.invoke(urls, "objectAtIndex:", i)
-                            val nsPath = Foundation.invoke(url, "path")
-                            runCatching { Foundation.toStringViaUTF8(nsPath) }.getOrNull()?.let(::File)
-                        }
+    private suspend fun pickFilesAndFolders(): List<File> = withContext(Dispatchers.IO) {
+        val pool = Foundation.NSAutoreleasePool()
+        try {
+            var result: List<File> = emptyList()
+            Foundation.executeOnMainThread(withAutoreleasePool = true, waitUntilDone = true) {
+                val panel = Foundation.invoke("NSOpenPanel", "openPanel")
+                Foundation.invoke(panel, "setCanChooseFiles:", true)
+                Foundation.invoke(panel, "setCanChooseDirectories:", true)
+                Foundation.invoke(panel, "setAllowsMultipleSelection:", true)
+                val response = Foundation.invoke(panel, "runModal")
+                if (response.toInt() == NS_MODAL_RESPONSE_OK) {
+                    val urls = Foundation.invoke(panel, "URLs")
+                    val count = Foundation.invoke(urls, "count").toInt()
+                    result = (0 until count).mapNotNull { i ->
+                        val url = Foundation.invoke(urls, "objectAtIndex:", i)
+                        val nsPath = Foundation.invoke(url, "path")
+                        runCatching { Foundation.toStringViaUTF8(nsPath) }.getOrNull()?.let(::File)
                     }
                 }
-                result
-            } finally {
-                pool.drain()
             }
+            result
+        } finally {
+            pool.drain()
         }
+    }
+
+    private companion object {
+        private const val NS_MODAL_RESPONSE_OK = 1
     }
 }
